# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zero Club API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{userId}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{userId}/update:
    put:
      operationId: updateProfile
      tags: [profiles]
      summary: Update own profile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List community feed posts
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: track
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Toggle like on a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /posts/{postId}/bookmark:
    post:
      operationId: bookmarkPost
      tags: [posts]
      summary: Toggle bookmark on a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkResponse"
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [posts]
      summary: List comments on a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [posts]
      summary: Create a comment
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get current user XP wallet
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/events:
    get:
      operationId: listXpEvents
      tags: [wallet]
      summary: Get XP event history
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/XpEvent"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        track:
          type: string
        level:
          type: integer
        xpBalance:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - displayName
        - track
        - level
        - xpBalance
        - createdAt
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        track:
          type: string
    Post:
      type: object
      properties:
        id:
          type: string
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/Profile"
        body:
          type: string
        imageUrl:
          type: string
          nullable: true
        track:
          type: string
        isProofProject:
          type: boolean
        xpAwarded:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        isLiked:
          type: boolean
        isBookmarked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorId
        - author
        - body
        - track
        - isProofProject
        - xpAwarded
        - likeCount
        - commentCount
        - isLiked
        - isBookmarked
        - createdAt
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - posts
        - total
        - hasMore
    CreatePostRequest:
      type: object
      properties:
        body:
          type: string
          maxLength: 500
        imageUrl:
          type: string
          nullable: true
        track:
          type: string
          enum: [product_design, frontend, growth, branding, mentorship]
        isProofProject:
          type: boolean
      required:
        - body
        - track
        - isProofProject
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
      required:
        - liked
        - likeCount
    BookmarkResponse:
      type: object
      properties:
        bookmarked:
          type: boolean
      required:
        - bookmarked
    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/Profile"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - author
        - body
        - createdAt
    CreateCommentRequest:
      type: object
      properties:
        body:
          type: string
      required:
        - body
...
    Wallet:
      type: object
      properties:
        userId:
          type: string
        xpBalance:
          type: integer
        level:
          type: integer
        xpToNextLevel:
          type: integer
        xpForCurrentLevel:
          type: integer
        totalXpForNextLevel:
          type: integer
      required:
        - userId
        - xpBalance
        - level
        - xpToNextLevel
        - xpForCurrentLevel
        - totalXpForNextLevel
    XpEvent:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        source:
          type: string
          enum: [build_posted, proof_project, bootcamp_module, bootcamp_completed]
        detail:
          type: string
          nullable: true
        amount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - source
        - amount
        - createdAt
...